ado_template_tracker/requirements.txt
pydantic>=2.10.6
requests>=2.31.0
azure-identity>=1.15.0
pyyaml>=6.0.0
azure-devops>=7.1.0b3
python-dotenv>=1.0.0
aiohttp>=3.8.1
rich>=10.12.0
jsonschema>=4.23.0
